Code BE : milestone eval form

+++ BE/online-learning-system/src/main/java/swp490/g23/onlinelearningsystem/entities/eval_detail/domain/EvalDetail.java
@@ -25,9 +25,9 @@ public class EvalDetail extends BaseEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long evalDetailId;
 
-    private Integer totalLoc;
+    private Long totalLoc;
 
-    private Long grade;
+    private Double grade;
 
     private String comment;
